Split things into separate class files, move entirely to singletons for clients
