+++ src/clients/db.ts
@@ -1,6 +1,6 @@
 import { Pool } from "pg";
 
-export const pool = new Pool({
+export const databaseClient = new Pool({
   connectionString: process.env.DATABASE_URL,
   ssl: {
     rejectUnauthorized: false,
