+++ src/commands/clan/_clans.ts
