+++ src/commands/clan/clan.ts
@@ -6,10 +6,10 @@ import {
   underscore,
 } from "discord.js";
 
-import { DREAD_CLANS } from "../../clans";
-import { createEmbed } from "../../discord";
-import { DetailedDreadStatus, client } from "../../kol";
+import { createEmbed } from "../../clients/discord";
+import { DetailedDreadStatus, kolClient } from "../../clients/kol";
 import { pluralize } from "../../utils";
+import { DREAD_CLANS } from "./_clans";
 
 const DREAD_BOSS_MAPPINGS = new Map([
   ["werewolf", "Air Wolf"],
@@ -165,7 +165,7 @@ export async function execute(interaction: ChatInputCommandInteraction) {
   }
   await interaction.deferReply();
   try {
-    const status = await client.getDetailedDreadStatus(clan.id);
+    const status = await kolClient.getDetailedDreadStatus(clan.id);
     const embed = createEmbed().setTitle(`Status update for ${clan.name}`);
 
     embed.setDescription(
