+++ src/commands/clan/status.ts
@@ -1,9 +1,9 @@
 import { ChatInputCommandInteraction, SlashCommandBuilder, bold } from "discord.js";
 
-import { DREAD_CLANS, clanState } from "../../clans";
-import { pool } from "../../db";
-import { client } from "../../kol";
+import { databaseClient } from "../../clients/db";
+import { kolClient } from "../../clients/kol";
 import { pluralize } from "../../utils";
+import { DREAD_CLANS, clanState } from "./_clans";
 
 export async function execute(interaction: ChatInputCommandInteraction) {
   await interaction.deferReply();
@@ -11,7 +11,7 @@ export async function execute(interaction: ChatInputCommandInteraction) {
   try {
     const dreadStatus = await Promise.all(
       DREAD_CLANS.map(async (clan) => {
-        const overview = await client.getDreadStatusOverview(clan.id);
+        const overview = await kolClient.getDreadStatusOverview(clan.id);
 
         const skills = overview.castle ? overview.skills : 0;
 
@@ -46,11 +46,11 @@ export const data = new SlashCommandBuilder()
   .setDescription("Get the current status of all monitored Dreadsylvania instances.");
 
 export async function async() {
-  clanState.parsedRaids = (await pool.query("SELECT raid_id FROM tracked_instances;")).rows.map(
-    (row) => row.raid_id
-  );
+  clanState.parsedRaids = (
+    await databaseClient.query("SELECT raid_id FROM tracked_instances;")
+  ).rows.map((row) => row.raid_id);
 
-  for (let player of (await pool.query("SELECT * FROM players;")).rows) {
+  for (let player of (await databaseClient.query("SELECT * FROM players;")).rows) {
     clanState.killMap.set(player.username, {
       kills: player.kills,
       skills: player.skills,
