--- src/things.ts
@@ -1,1127 +1,0 @@
-import { EmbedBuilder, bold, hyperlink } from "discord.js";
-
-import { KoLClient } from "./kol";
-import { cleanString, indent, pluralize, toWikiLink } from "./utils";
-
-type FamiliarActionTypes =
-  | "none"
-  | "stat0"
-  | "stat1"
-  | "item0"
-  | "meat0"
-  | "combat0"
-  | "combat1"
-  | "drop"
-  | "block"
-  | "delevel"
-  | "hp0"
-  | "mp0"
-  | "meat1"
-  | "stat2"
-  | "other0"
-  | "hp1"
-  | "mp1"
-  | "stat3"
-  | "other1"
-  | "passive"
-  | "underwater"
-  | "variable";
-
-type FamiliarClassification = {
-  combination: FamiliarActionTypes[];
-  description: string;
-};
-
-const FAMILIAR_CLASSIFCATIONS: FamiliarClassification[] = [
-  {
-    combination: ["combat0", "meat1", "delevel", "hp0", "mp0"],
-    description: "Cocoabo-like.",
-  },
-  {
-    combination: ["combat0", "combat1", "mp0", "hp0"],
-    description: "Deals elemental and physical damage to restore your hp and mp in combat.",
-  },
-  {
-    combination: ["combat0", "combat1", "mp0"],
-    description: "Deals elemental and physical damage to restore your mp in combat.",
-  },
-  {
-    combination: ["combat0", "combat1", "hp0"],
-    description: "Deals elemental and physical damage to heal you in combat.",
-  },
-  {
-    combination: ["combat0", "mp0", "hp0"],
-    description: "Deals physical damage to restore your hp and mp in combat.",
-  },
-  {
-    combination: ["combat0", "mp0"],
-    description: "Deals physical damage to restore your mp in combat.",
-  },
-  {
-    combination: ["combat0", "hp0"],
-    description: "Deals physical damage to heal you in combat.",
-  },
-  {
-    combination: ["combat1", "mp0", "hp0"],
-    description: "Deals elemental damage to restore your hp and mp in combat.",
-  },
-  {
-    combination: ["combat1", "mp0"],
-    description: "Deals elemental damage to restore your mp in combat.",
-  },
-  {
-    combination: ["combat1", "hp0"],
-    description: "Deals elemental damage to heal you in combat.",
-  },
-  {
-    combination: ["combat0", "combat1"],
-    description: "Deals physical and elemental damage in combat.",
-  },
-  {
-    combination: ["combat0"],
-    description: "Deals physical damage in combat.",
-  },
-  {
-    combination: ["combat1"],
-    description: "Deals elemental damage in combat.",
-  },
-  {
-    combination: ["item0", "meat0", "stat0"],
-    description: "Boosts item drops, meat drops and stat gains (volleyball-like).",
-  },
-  {
-    combination: ["meat0", "stat0"],
-    description: "Boosts meat drops and stat gains (volleyball-like).",
-  },
-  {
-    combination: ["item0", "stat0"],
-    description: "Boosts item drops and stat gains (volleyball-like).",
-  },
-  {
-    combination: ["item0", "meat0", "stat1"],
-    description: "Boosts item drops, meat drops and stat gains (sombrero-like).",
-  },
-  {
-    combination: ["meat0", "stat1"],
-    description: "Boosts meat drops and stat gains (sombrero-like).",
-  },
-  {
-    combination: ["item0", "stat1"],
-    description: "Boosts item drops and stat gains (sombrero-like).",
-  },
-  {
-    combination: ["stat0"],
-    description: "Boosts stat gains (volleyball-like).",
-  },
-  {
-    combination: ["stat1"],
-    description: "Boosts stat gains (sombero-like).",
-  },
-  {
-    combination: ["meat0", "item0"],
-    description: "Boosts item and meat drops.",
-  },
-  {
-    combination: ["item0"],
-    description: "Boosts item drops.",
-  },
-  {
-    combination: ["meat0"],
-    description: "Boosts meat drops",
-  },
-  {
-    combination: ["block"],
-    description: "Staggers enemies in combat.",
-  },
-  {
-    combination: ["delevel"],
-    description: "Delevels enemies in combat.",
-  },
-  {
-    combination: ["hp0", "mp0"],
-    description: "Restores your hp and mp during combat.",
-  },
-  {
-    combination: ["hp0"],
-    description: "Heals you during combat.",
-  },
-  {
-    combination: ["mp0"],
-    description: "Restores your mp during combat.",
-  },
-  {
-    combination: ["meat1"],
-    description: "Drops meat during combat.",
-  },
-  {
-    combination: ["stat2"],
-    description: "Grants stats during combat.",
-  },
-  {
-    combination: ["other0"],
-    description: "Does something unusual during combat.",
-  },
-  {
-    combination: ["hp1", "mp1"],
-    description: "Restores your hp and mp after combat.",
-  },
-  {
-    combination: ["hp1"],
-    description: "Heals you after combat.",
-  },
-  {
-    combination: ["mp1"],
-    description: "Restores mp after combat.",
-  },
-  {
-    combination: ["stat3"],
-    description: "Grants stats after combat.",
-  },
-  {
-    combination: ["other1"],
-    description: "Does something unusual after combat.",
-  },
-  {
-    combination: ["passive"],
-    description: "Grants a passive benefit.",
-  },
-  {
-    combination: ["drop"],
-    description: "Drops special items.",
-  },
-  {
-    combination: ["variable"],
-    description: "Has varying abilities.",
-  },
-  {
-    combination: ["none"],
-    description: "Does nothing useful.",
-  },
-  {
-    combination: ["underwater"],
-    description: "Can naturally breathe underwater.",
-  },
-];
-
-const HARD_CODED_FAMILIARS: Map<string, string> = new Map([
-  ["baby mutant rattlesnake", "Boosts stat gains based on Grimace Darkness.\n"],
-  ["chocolate lab", "Boosts item and sprinkle drops.\n"],
-  ["cute meteor", "Increases your combat initiative.\nDeals elemental damage in combat.\n"],
-  ["disgeist", "Decreases your combat frequency.\n"],
-  ["exotic parrot", "Increases your elemental resistances.\n"],
-  [
-    "happy medium",
-    "Increases your combat initiative.\nBoosts stat gains (volleyball-like).\nDrops special items.\n",
-  ],
-  [
-    "god lobster",
-    "Boosts stat gains (volleyball-like).\nHas varying abilities.\nCan naturally breathe underwater.\n",
-  ],
-  ["hobo monkey", "Massively boosts meat drops.\nDrops meat during combat.\n"],
-  ["jumpsuited hound dog", "Massively boosts item drops.\nIncreases your combat frequency.\n"],
-  ["magic dragonfish", "Increases your spell damage.\n"],
-  ["peppermint rhino", "Boosts item drops, especially in Dreadsylvania.\n"],
-  ["melodramedary", "Boosts stat gains (volleyball-like).\nRestores your mp after combat.\n"],
-  ["mu", "Increases your elemental resistances.\nDeals elemental damage in combat.\n"],
-  ["mutant cactus bud", "Boosts meat drops based on Grimace Darkness.\n"],
-  ["mutant fire ant", "Boosts item drops based on Grimace Darkness.\n"],
-  ["oily woim", "Increases your combat initiative.\n"],
-  ["peppermint rhino", "Boosts item drops, especially candy drops.\n"],
-  ["purse rat", "Increases your monster level.\n"],
-  ["red-nosed snapper", "Boosts item drops, especially underwater.\nDrops special items.\n"],
-  ["robortender", "Boosts your meat drops.\nDrops special items.\nHas varying abilities.\n"],
-  [
-    "space jellyfish",
-    "Increases your combat initiative.\nBoosts your item drops and stat gains (volleyball-like), especially underwater.\nDelevels enemies in combat.\nDrops special items.\n",
-  ],
-  [
-    "steam-powered cheerleader",
-    "Boosts item drops based on remaining steam.\nDelevels enemies in combat based on remaining steam.\nDoes something unusual during combat.\n",
-  ],
-  ["trick-or-treating tot", "Restores your hp and mp after combat.\nHas varying abilities.\n"],
-  ["xiblaxian holo-companion", "Increases your combat initiative.\nStaggers enemies in combat.\n"],
-  ["black cat", "Is adorable.\nGenerally messes with you.\n"],
-  ["o.a.f.", "Is optimal.\nGenerally messes with you.\n"],
-]);
-
-export abstract class Thing {
-  abstract name(): string;
-  abstract addToEmbed(embed: EmbedBuilder, client: KoLClient): Promise<void>;
-}
-
-type ItemData = {
-  id: number;
-  name: string;
-  descId: number;
-  imageUrl: string;
-  types: string[];
-  quest: boolean;
-  gift: boolean;
-  tradeable: boolean;
-  discardable: boolean;
-  autosell: number;
-  pluralName: string;
-};
-
-export class Item implements Thing {
-  _item: ItemData;
-  _name: string;
-  _shortDescription: string = "";
-  _addlDescription: string = "";
-  _fullDescription: string = "";
-  _craftSting: string = "";
-  _container: Item | undefined;
-  _contents: Item | undefined;
-  _zapGroup: Item[] | undefined;
-  _foldGroup: Item[] | undefined;
-
-  constructor(data: string, itemMap: Map<string, string[]>) {
-    this._item = this.parseItemData(data);
-    this._name = this._item.name.toLowerCase();
-    this._shortDescription = this.buildShortDescription(itemMap);
-  }
-
-  mapQuality(rawQuality: string): string {
-    switch (rawQuality) {
-      case "crappy":
-        return "Crappy";
-      case "decent":
-        return "Decent";
-      case "good":
-        return "Good";
-      case "awesome":
-        return "Awesome";
-      case "EPIC":
-        return "EPIC";
-      default:
-        return "????";
-    }
-  }
-
-  mapStat(rawStat: string): string {
-    switch (rawStat) {
-      case "Mys":
-        return "Mysticality";
-      case "Mox":
-        return "Moxie";
-      case "Mus":
-        return "Muscle";
-      default:
-        return "???";
-    }
-  }
-
-  pluralizeAdventures(range: string) {
-    const usePlural = range !== "1";
-    return `${range} adventure${usePlural ? "s" : ""}`;
-  }
-
-  buildShortDescription(itemMap: Map<string, string[]>): string {
-    let tradeability_section = "";
-    if (this._item.quest) tradeability_section += "Quest Item\n";
-    else if (this._item.gift) tradeability_section += "Gift Item\n";
-    if (!this._item.tradeable) {
-      if (this._item.discardable) tradeability_section += "Cannot be traded.\n";
-      else tradeability_section += "Cannot be traded or discarded.\n";
-    } else if (!this._item.discardable) tradeability_section += "Cannot be discarded.\n";
-
-    const data = itemMap.get(this._name) || [];
-    let desc = `(Item ${this._item.id})\n`;
-
-    if (this._item.types.includes("food")) {
-      const advRange = data[4].split("-");
-      const size = parseInt(data[1]);
-      const average =
-        advRange.length > 1
-          ? (parseInt(advRange[0]) + parseInt(advRange[1])) / 2
-          : parseInt(data[4]);
-      desc += `${bold(`${this.mapQuality(data[3])} food`)} (Size ${data[1]}${
-        data[2] !== "1" ? `, requires level ${data[2]}` : ""
-      })`;
-      if (data[4] !== "0") {
-        desc += `\n${this.pluralizeAdventures(data[4])}`;
-        if (advRange.length > 1 || size > 1) {
-          desc += ` (${advRange.length > 1 ? `Average ${pluralize(average, "adventure")}` : ""}`;
-          desc += `${
-            size > 1
-              ? `${advRange.length > 1 ? ", " : ""}${Number(
-                  (average / size).toFixed(2)
-                )} per fullness`
-              : ""
-          })`;
-        }
-      }
-      desc += `\n${tradeability_section}\n`;
-      return desc;
-    }
-    if (this._item.types.includes("drink")) {
-      const advRange = data[4].split("-");
-      const size = parseInt(data[1]);
-      const average =
-        advRange.length > 1
-          ? (parseInt(advRange[0]) + parseInt(advRange[1])) / 2
-          : parseInt(data[4]);
-      desc += `${bold(`${this.mapQuality(data[3])} booze`)} (Potency ${data[1]}${
-        data[2] !== "1" ? `, requires level ${data[2]}` : ""
-      })`;
-      if (data[4] !== "0") {
-        desc += `\n${this.pluralizeAdventures(data[4])}`;
-        if (advRange.length > 1 || size > 1) {
-          desc += ` (${advRange.length > 1 ? `Average ${pluralize(average, "adventure")}` : ""}`;
-          desc += `${
-            size > 1
-              ? `${advRange.length > 1 ? ", " : ""}${Number(
-                  (average / size).toFixed(2)
-                )} per inebriety`
-              : ""
-          })`;
-        }
-      }
-      desc += `\n${tradeability_section}\n`;
-      return desc;
-    }
-    if (this._item.types.includes("spleen")) {
-      const advRange = data[4].split("-");
-      const size = parseInt(data[1]);
-      const average =
-        advRange.length > 1
-          ? (parseInt(advRange[0]) + parseInt(advRange[1])) / 2
-          : parseInt(data[4]);
-      desc += `${bold(`${this.mapQuality(data[3])} spleen item`)} (Toxicity ${data[1]}${
-        data[2] !== "1" ? `, requires level ${data[2]}` : ""
-      })`;
-      if (data[4] !== "0") {
-        desc += `\n${this.pluralizeAdventures(data[4])}`;
-        if (advRange.length > 1 || size > 1) {
-          desc += ` (${advRange.length > 1 ? `Average ${pluralize(average, "adventure")}` : ""}`;
-          desc += `${
-            size > 1
-              ? `${advRange.length > 1 ? ", " : ""}${Number(
-                  (average / size).toFixed(2)
-                )} per spleen`
-              : ""
-          })`;
-        }
-      }
-      desc += `\n${tradeability_section}\n`;
-      return desc;
-    }
-    if (this._item.types.includes("weapon")) {
-      const requirement = (data[2] || ": 0").split(": ");
-      const damage = parseInt(data[1]) / 10;
-      let equipString = `${bold(data[3])}\n`;
-      equipString += `${Math.round(damage)}-${Math.round(damage * 2)} damage${
-        data[2] !== "none" && requirement[1] !== "0"
-          ? `, requires ${requirement[1]} ${this.mapStat(requirement[0])}`
-          : ""
-      }\n`;
-      equipString += `${tradeability_section}\n`;
-      return equipString;
-    }
-    if (this._item.types.includes("offhand")) {
-      const requirement = (data[2] || ": 0").split(": ");
-      const isShield = data[3] === "shield";
-      let equipString = bold(`Offhand ${isShield ? "Shield" : "Item"}`);
-      equipString += ` (${data[1]} power${
-        data[2] !== "none" && requirement[1] !== "0"
-          ? `, requires ${requirement[1]} ${this.mapStat(requirement[0])}`
-          : ""
-      })\n`;
-      if (isShield) equipString += `Damage Reduction: ${parseInt(data[1]) / 15 - 1}\n`;
-      equipString += `${tradeability_section}\n`;
-      return equipString;
-    }
-    if (this._item.types.includes("container")) {
-      const requirement = (data[2] || ": 0").split(": ");
-      let equipString = bold(`Back Item`);
-      equipString += ` (${data[1]} power${
-        data[2] !== "none" && requirement[1] !== "0"
-          ? `, requires ${requirement[1]} ${this.mapStat(requirement[0])}`
-          : ""
-      })\n`;
-      equipString += `${tradeability_section}\n`;
-      return equipString;
-    }
-    if (this._item.types.includes("familiar")) {
-      return `${bold("Familiar Equipment")}\n`;
-    }
-    //This is hideous. but hey.
-    let equipmentType = "";
-    for (equipmentType of ["hat", "shirt", "pants", "accessory", ""]) {
-      if (this._item.types.includes(equipmentType)) break;
-    }
-    if (equipmentType) {
-      const requirement = (data[2] || ": 0").split(": ");
-      let equipString = bold(equipmentType.charAt(0).toUpperCase() + equipmentType.slice(1));
-      equipString += ` (${data[1]} power${
-        data[2] !== "none" && requirement[1] !== "0"
-          ? `, requires ${requirement[1]} ${this.mapStat(requirement[0])}`
-          : ""
-      })\n`;
-      equipString += `${tradeability_section}\n`;
-      return equipString;
-    }
-    if (this._item.types.includes("potion")) {
-      if (this._item.types.includes("combat")) return `${bold("Potion")} (also usable in combat)\n`;
-      return `${bold("Potion")}\n${tradeability_section}`;
-    }
-    if (this._item.types.includes("reusable")) {
-      if (this._item.types.includes("combat"))
-        return `${bold("Reusable item")} (also usable in combat)\n`;
-      return `${bold("Reusable item")}\n${tradeability_section}`;
-    }
-    if (this._item.types.includes("usable") || this._item.types.includes("multiple")) {
-      if (this._item.types.includes("combat"))
-        return `${bold("Usable item")} (also usable in combat)\n`;
-      return `${bold("Usable item")}\n${tradeability_section}`;
-    }
-    if (this._item.types.includes("combat")) {
-      return `${bold("Combat item")}\n${tradeability_section}`;
-    }
-    if (this._item.types.includes("combat reusable")) {
-      return `${bold("Reusable combat item")}\n${tradeability_section}`;
-    }
-    if (this._item.types.includes("grow")) {
-      return `${bold("Familiar hatchling")}\n${tradeability_section}`;
-    }
-    return `${bold("Miscellaneous Item")}\n${tradeability_section}`;
-  }
-
-  addGrowingFamiliar(familiar: Familiar): void {
-    this._addlDescription = `Grows into: ${bold(
-      hyperlink(familiar.get().name, toWikiLink(familiar.get().name))
-    )}\n`;
-  }
-
-  addEquppingFamiliar(familiar: Familiar): void {
-    this._addlDescription = `Familiar: ${hyperlink(
-      familiar.get().name,
-      toWikiLink(familiar.get().name)
-    )}\n`;
-  }
-
-  addContainer(container: Item) {
-    this._container = container;
-  }
-
-  addContents(contents: Item) {
-    this._contents = contents;
-  }
-
-  addZapGroup(zapGroup: Item[]) {
-    this._zapGroup = zapGroup;
-  }
-
-  addFoldGroup(foldGroup: Item[]) {
-    this._foldGroup = foldGroup;
-  }
-
-  get(): ItemData {
-    return this._item;
-  }
-
-  name(): string {
-    return this._name;
-  }
-
-  async getMallPrice(client: KoLClient) {
-    if (!this._item.tradeable) return "";
-
-    const { mallPrice, limitedMallPrice, formattedMallPrice, formattedLimitedMallPrice } =
-      await client.getMallPrice(this._item.id);
-
-    const url = `https://g1wjmf0i0h.execute-api.us-east-2.amazonaws.com/default/itemgraph?itemid=${this._item.id}&timespan=1&noanim=0`;
-
-    if (mallPrice) {
-      let output = `Mall Price: ${hyperlink(`${formattedMallPrice} meat`, url)}`;
-      if (limitedMallPrice && limitedMallPrice < mallPrice) {
-        output += ` (or ${formattedLimitedMallPrice} meat limited per day)`;
-      }
-      return output;
-    } else if (limitedMallPrice) {
-      return `Mall Price: ${hyperlink(
-        `${formattedLimitedMallPrice} meat (only available limited per day)`,
-        url
-      )}`;
-    } else {
-      return "Mall extinct.";
-    }
-  }
-
-  async buildFullDescription(client: KoLClient, withAddl: boolean = true): Promise<string> {
-    let description_string = this._shortDescription + (withAddl ? this._addlDescription : "");
-
-    let blueText = await client.getItemDescription(this._item.descId);
-
-    let autosell = "";
-    if (this._item.discardable && this._item.autosell > 0) {
-      autosell += `Autosell value: ${this._item.autosell} meat.`;
-    }
-
-    let price_section = await this.getMallPrice(client);
-
-    let container = "";
-    if (withAddl && this._container) {
-      container = `\nEnclosed in: ${bold(
-        hyperlink(this._container.get().name, toWikiLink(this._container.get().name))
-      )}\n${await this._container?.buildFullDescription(client, false)}\n`;
-    }
-
-    let contents = "";
-    if (withAddl && this._contents) {
-      contents = `\nEncloses: ${bold(
-        hyperlink(this._contents.get().name, toWikiLink(this._contents.get().name))
-      )}\n${await this._contents?.buildFullDescription(client, false)}\n`;
-    }
-
-    let zapGroup = "";
-    if (this._zapGroup) {
-      zapGroup = `\nZaps into: ${this._zapGroup
-        .filter((item) => item.name() !== this.name())
-        .slice(0, 7)
-        .map((item) => item.get().name)
-        .map((name) => hyperlink(name, toWikiLink(name)))
-        .join(", ")}${this._zapGroup.length > 8 ? " ...and more." : ""}\n`;
-
-      const tradeables = (
-        await Promise.all(
-          this._zapGroup
-            .filter((item) => item.get().tradeable)
-            .map(async (item) => ({ item: item, price: await client.getMallPrice(item.get().id) }))
-        )
-      )
-        .filter((item) => item.price.minPrice)
-        .sort((a, b) => (a.price.minPrice as number) - (b.price.minPrice as number));
-      if (tradeables.length > 0) {
-        if (tradeables[0].item.get().id === this._item.id) {
-          zapGroup += "(This item is the cheapest in its zap group)";
-        } else {
-          const cheapest = tradeables[0].item;
-          zapGroup += `(Cheapest: ${hyperlink(
-            cheapest.get().name,
-            toWikiLink(cheapest.get().name)
-          )} @ ${hyperlink(
-            `${tradeables[0].price.formattedMinPrice} meat`,
-            `https://g1wjmf0i0h.execute-api.us-east-2.amazonaws.com/default/itemgraph?itemid=${
-              cheapest.get().id
-            }&timespan=1&noanim=0`
-          )})\n`;
-        }
-      }
-    }
-
-    let foldGroup = "";
-    if (this._foldGroup) {
-      foldGroup = `\nFolds into: ${this._foldGroup
-        .filter((item) => item.name() !== this.name())
-        .slice(0, 7)
-        .map((item) => item.get().name)
-        .map((name) => hyperlink(name, toWikiLink(name)))
-        .join(", ")}${this._foldGroup.length > 8 ? " ...and more." : ""}\n`;
-
-      const tradeables = (
-        await Promise.all(
-          this._foldGroup
-            .filter((item) => item.get().tradeable)
-            .map(async (item) => ({ item: item, price: await client.getMallPrice(item.get().id) }))
-        )
-      )
-        .filter((item) => item.price.minPrice)
-        .sort((a, b) => (a.price.minPrice as number) - (b.price.minPrice as number));
-      if (tradeables.length > 0) {
-        if (tradeables[0].item.get().id === this._item.id) {
-          foldGroup += "(This item is the cheapest in its fold group)";
-        } else {
-          const cheapest = tradeables[0].item;
-          foldGroup += `(Cheapest: ${hyperlink(
-            cheapest.get().name,
-            toWikiLink(cheapest.get().name)
-          )}) @ ${hyperlink(
-            `${tradeables[0].price.formattedMinPrice} meat`,
-            `https://g1wjmf0i0h.execute-api.us-east-2.amazonaws.com/default/itemgraph?itemid=${
-              cheapest.get().id
-            }&timespan=1&noanim=0`
-          )})\n`;
-        }
-      }
-    }
-
-    if (blueText && (autosell || price_section)) blueText += "\n";
-    if (autosell) autosell += "\n";
-    if (price_section) price_section += "\n";
-
-    return `${description_string}${blueText}${autosell}${price_section}${zapGroup}${foldGroup}${container}${contents}`;
-  }
-
-  async addToEmbed(embed: EmbedBuilder, client: KoLClient): Promise<void> {
-    embed.setThumbnail(`http://images.kingdomofloathing.com/itemimages/${this._item.imageUrl}`);
-    embed.setDescription(await this.buildFullDescription(client));
-  }
-
-  parseItemData(itemData: string): ItemData {
-    const data = itemData.split(/\t/);
-    if (data.length < 7) {
-      throw "Invalid data";
-    }
-    return {
-      id: parseInt(data[0]),
-      name: cleanString(data[1]),
-      descId: parseInt(data[2]),
-      imageUrl: data[3],
-      types: data[4].split(", "),
-      quest: data[5].indexOf("q") >= 0,
-      gift: data[5].indexOf("g") >= 0,
-      tradeable: data[5].indexOf("t") >= 0,
-      discardable: data[5].indexOf("d") >= 0,
-      autosell: parseInt(data[6]),
-      pluralName: cleanString(data[7]) || `${cleanString(data[1])}s`,
-    };
-  }
-}
-
-type EffectData = {
-  id: number;
-  name: string;
-  imageUrl: string;
-  descId: string;
-  quality: string;
-  hookah: boolean;
-};
-
-type PizzaData = {
-  letters: string;
-  options: number;
-};
-
-export class Effect implements Thing {
-  _effect: EffectData;
-  _name: string;
-  _pizza?: PizzaData;
-
-  constructor(data: string, avatarSet: Set<string>) {
-    this._effect = this.parseEffectData(data, avatarSet);
-    this._name = this._effect.name.toLowerCase();
-  }
-
-  get(): EffectData {
-    return this._effect;
-  }
-
-  name(): string {
-    return this._name;
-  }
-
-  async addToEmbed(embed: EmbedBuilder, client: KoLClient): Promise<void> {
-    embed.setThumbnail(`http://images.kingdomofloathing.com/itemimages/${this._effect.imageUrl}`);
-    let description = `${bold("Effect")}\n(Effect ${
-      this.get().id
-    })\n${await client.getEffectDescription(this._effect.descId)}\n\n`;
-    if (this._effect.hookah) {
-      if (this._pizza) {
-        description += `Pizza: ${this._pizza.letters.padEnd(4, "✱")} (${
-          this._pizza.options === 1 ? "Uncontested" : `1 in ${this._pizza.options}`
-        })`;
-      } else {
-        description +=
-          "Pizza: If you are reading this, Captain 'Jalen' Scotch has hecked something up. Please ping him.";
-      }
-    } else {
-      description += "Inelgible for pizza, wishes, or hookahs.";
-    }
-    embed.setDescription(description);
-  }
-
-  parseEffectData(effectData: string, avatarPotionSet: Set<string>): EffectData {
-    const data = effectData.split(/\t/);
-    if (data.length < 6) throw "Invalid data";
-    const potion = data.length >= 7 ? (data[6].startsWith("use 1") ? data[6].slice(6) : "") : "";
-    const isAvatar = avatarPotionSet.has(potion.toLowerCase());
-    if (isAvatar) avatarPotionSet.delete(potion.toLowerCase());
-    return {
-      id: parseInt(data[0]),
-      name: cleanString(data[1]),
-      imageUrl: data[2],
-      descId: data[3],
-      quality: data[4],
-      hookah: !isAvatar && data[5].indexOf("nohookah") === -1 && data[4] !== "bad",
-    };
-  }
-}
-
-type SkillData = {
-  id: number;
-  name: string;
-  imageUrl: string;
-  type: number;
-  manaCost: number;
-  duration: number;
-  level?: number;
-};
-
-export class Skill implements Thing {
-  _skill: SkillData;
-  _name: string;
-  _description: string = "";
-
-  constructor(data: string) {
-    this._skill = this.parseSkillData(data);
-    this._name = this._skill.name.toLowerCase();
-  }
-
-  get(): SkillData {
-    return this._skill;
-  }
-
-  name(): string {
-    return this._name;
-  }
-
-  block(): number {
-    return Math.floor(this._skill.id / 1000);
-  }
-
-  async buildDescription(client: KoLClient): Promise<string> {
-    let description = `(Skill ${this._skill.id})\n`;
-    switch (this._skill.type) {
-      case 0:
-        description += bold("Passive Skill");
-        break;
-      case 1:
-      case 2:
-      case 3:
-      case 4:
-        description += `${bold("Skill")}\nCost: ${this._skill.manaCost}mp`;
-        break;
-      case 5:
-        description += `${bold("Combat Skill")}\nCost: ${this._skill.manaCost}mp`;
-        break;
-      case 6:
-        description += `${bold("Skill (Boris Song)")}\nCost: ${this._skill.manaCost}mp`;
-        break;
-      case 7:
-        description += `${bold("Combat/Noncombat Skill")}\nCost: ${this._skill.manaCost}mp`;
-        break;
-      case 8:
-        description += bold("Combat Passive Skill");
-        break;
-      case 9:
-        description += `${bold("Skill (Expression)")}\nCost: ${this._skill.manaCost}mp`;
-        break;
-      case 10:
-        description += `${bold("Skill (Walk)")}\nCost: ${this._skill.manaCost}mp`;
-        break;
-      default:
-        description += bold("Skill");
-        break;
-    }
-    description += "\n\n";
-    description += await client.getSkillDescription(this._skill.id);
-    return description;
-  }
-
-  async addToEmbed(embed: EmbedBuilder, client: KoLClient): Promise<void> {
-    embed.setThumbnail(`http://images.kingdomofloathing.com/itemimages/${this._skill.imageUrl}`);
-    if (!this._description) this._description = await this.buildDescription(client);
-    embed.setDescription(this._description);
-  }
-
-  parseSkillData(skillData: string): SkillData {
-    const data = skillData.split(/\t/);
-    if (data.length < 6) throw "Invalid data";
-    return {
-      id: parseInt(data[0]),
-      name: cleanString(data[1]),
-      imageUrl: data[2],
-      type: parseInt(data[3]),
-      manaCost: parseInt(data[4]),
-      duration: parseInt(data[5]),
-      level: data[6] ? parseInt(data[6]) : undefined,
-    };
-  }
-}
-
-type FamiliarData = {
-  id: number;
-  name: string;
-  imageUrl: string;
-  types: string[];
-  larva: string;
-  item: string;
-  attributes: string;
-};
-
-export class Familiar implements Thing {
-  _familiar: FamiliarData;
-  _name: string;
-  _description: string = "";
-  _hatchling: Item | undefined;
-  _equipment: Item | undefined;
-  _typeString: string = "";
-
-  constructor(data: string) {
-    this._familiar = this.parseFamiliarData(data);
-    this._name = this._familiar.name.toLowerCase();
-  }
-
-  get(): FamiliarData {
-    return this._familiar;
-  }
-
-  name(): string {
-    return this._name;
-  }
-
-  addHatchling(hatchling: Item) {
-    this._hatchling = hatchling;
-  }
-
-  addEquipment(equipment: Item) {
-    this._equipment = equipment;
-  }
-
-  parseTypes(): string {
-    if (!this._typeString) {
-      if (HARD_CODED_FAMILIARS.has(this.name()))
-        this._typeString = HARD_CODED_FAMILIARS.get(this.name()) || "";
-      else {
-        let types = "";
-        let typeArray = [...this._familiar.types] as FamiliarActionTypes[];
-
-        for (let classification of FAMILIAR_CLASSIFCATIONS) {
-          if (classification.combination.every((type) => typeArray.includes(type))) {
-            typeArray = typeArray.filter((type) => !classification.combination.includes(type));
-            types += `${classification.description}\n`;
-          }
-        }
-
-        this._typeString = types;
-      }
-    }
-    return this._typeString;
-  }
-
-  async buildDescription(client: KoLClient): Promise<string> {
-    let description_string = bold("Familiar") + "\n";
-    description_string += `${this.parseTypes()}\n`;
-    description_string += `Attributes: ${this._familiar.attributes || "None"}\n\n`;
-    if (this._familiar.larva)
-      description_string += `Hatchling: ${hyperlink(
-        this._familiar.larva,
-        toWikiLink(this._familiar.larva)
-      )}\n${(await this._hatchling?.buildFullDescription(client, false))
-        ?.substring(23)
-        .replace(/\n+/g, "\n")}\n`;
-    if (this._familiar.item)
-      description_string += `Equipment: ${hyperlink(
-        this._familiar.item,
-        toWikiLink(this._familiar.item)
-      )}\n${indent((await client.getItemDescription(this._equipment?.get().descId || 0)) || "")}`;
-    return description_string;
-  }
-
-  async addToEmbed(embed: EmbedBuilder, client: KoLClient): Promise<void> {
-    embed.setThumbnail(`http://images.kingdomofloathing.com/itemimages/${this._familiar.imageUrl}`);
-    if (!this._description) this._description = await this.buildDescription(client);
-    embed.setDescription(this._description);
-  }
-
-  parseFamiliarData(familiarData: string): FamiliarData {
-    const data = familiarData.split(/\t/);
-    if (data.length < 10) throw "Invalid data";
-    return {
-      id: parseInt(data[0]),
-      name: cleanString(data[1]),
-      imageUrl: data[2],
-      types: data[3].split(","),
-      larva: cleanString(data[4]),
-      item: cleanString(data[5]),
-      attributes: data[10] ? data[10].replace(/,/g, ", ") : "",
-    };
-  }
-}
-
-type Drop = {
-  item: string;
-  droprate: number;
-  attributes: {
-    pickpocketOnly: boolean;
-    noPickpocket: boolean;
-    conditional: boolean;
-    fixedRate: boolean;
-    accordion: boolean;
-  };
-};
-
-function convertToDrop(drop: string): Drop | undefined {
-  const dropMatch = drop.match(/^(?<itemName>.+) \((?<attributes>[a-z]*)(?<droprate>[\d]+)\)$/);
-  if (dropMatch) {
-    const attributes = dropMatch.groups?.attributes || "";
-    return {
-      item: dropMatch.groups?.itemName || "",
-      droprate: parseInt(dropMatch.groups?.droprate || "0"),
-      attributes: {
-        pickpocketOnly: attributes.indexOf("p") > -1,
-        noPickpocket: attributes.indexOf("n") > -1,
-        conditional: attributes.indexOf("c") > -1,
-        fixedRate: attributes.indexOf("f") > -1,
-        accordion: attributes.indexOf("a") > -1,
-      },
-    };
-  }
-  return undefined;
-}
-
-type MonsterData = {
-  name: string;
-  id: number;
-  imageUrl: string;
-  parameters: string;
-  drops: Drop[];
-};
-
-export class Monster implements Thing {
-  _monster: MonsterData;
-  _name: string;
-  _description: string = "";
-
-  constructor(data: string) {
-    this._monster = this.parseMonsterData(data);
-    this._name = this._monster.name.toLowerCase();
-  }
-
-  get(): MonsterData {
-    return this._monster;
-  }
-
-  name(): string {
-    return this._name;
-  }
-
-  async buildDescription(client: KoLClient): Promise<string> {
-    let description = `${bold("Monster")}\n(Monster ${this._monster.id})\n`;
-    const atk = this._monster.parameters.match(/Atk: (?<atk>\-?[\d]+)/);
-    const def = this._monster.parameters.match(/Def: (?<def>\-?[\d]+)/);
-    const hp = this._monster.parameters.match(/HP: (?<hp>\-?[\d]+)/);
-    const scale = this._monster.parameters.match(/Scale: (?<scale>-?[\d]+)/);
-    const floor = this._monster.parameters.match(/Floor: (?<floor>[\d]+)/);
-    const cap = this._monster.parameters.match(/Cap: (?<cap>[\d]+)/);
-
-    if (atk && def && hp) {
-      description += `Attack: ${atk[1]} | Defense: ${def[1]} | HP: ${hp[1]}\n`;
-    } else if (scale) {
-      let scaleString =
-        parseInt(scale[1]) >= 0 ? `plus ${scale[1]}` : `minus ${scale[1].substring(1)}`;
-      if (cap || floor) {
-        scaleString += " (";
-        if (floor) scaleString += `min ${floor[1]}`;
-        if (floor && cap) scaleString += ", ";
-        if (cap) scaleString += `max ${cap[1]}`;
-        scaleString += ")";
-      }
-      if (hp) {
-        description += `Scales to your stats ${scaleString} | HP: ${hp[1]}\n`;
-      } else {
-        description += `Scales to your stats ${scaleString} | HP: 75% of defense.\n`;
-      }
-    } else if (this._monster.parameters.indexOf("Scales: ")) {
-      let capString = "";
-      if (cap || floor) {
-        capString += " (";
-        if (floor) capString += `min ${floor[1]}`;
-        if (floor && cap) capString += ", ";
-        if (cap) capString += `max ${cap[1]}`;
-        capString += ")";
-      }
-      if (hp) description += `Scales to your stats${capString} | HP: ${hp[1]}\n`;
-      else description += `Scales to your stats${capString} | HP: 75% of defense.\n`;
-    } else description += "Scales unusually.\n";
-
-    const phylum = this._monster.parameters.match(/P: (?<phylum>[\a-z]+)/);
-    if (phylum) {
-      description += `Phylum: ${phylum[1]}\n`;
-    }
-    const element = this._monster.parameters.match(/E: (?<element>[\a-z]+)/);
-    if (element) {
-      description += `Element: ${element[1]}\n`;
-    }
-    if (this._monster.parameters.indexOf("Init: 10000") > -1)
-      description += "Always wins initiative.\n";
-    if (this._monster.parameters.indexOf("Init: -10000") > -1)
-      description += "Always loses initiative.\n";
-    if (this._monster.parameters.indexOf("FREE") > -1)
-      description += "Doesn't cost a turn to fight.\n";
-    if (this._monster.parameters.indexOf("NOCOPY") > -1) description += "Can't be copied.\n";
-    if (this._monster.parameters.indexOf("BOSS") > -1) description += "Instakill immune.\n";
-    if (this._monster.parameters.indexOf("ULTRARARE") > -1)
-      description += "Ultra-rare encounter.\n";
-
-    const meat = this._monster.parameters.match(/Meat: (?<meat>[\d]+)/);
-    if (this._monster.drops.length || meat) {
-      description += "\nDrops:\n";
-      if (meat) {
-        const meatInt = parseInt(meat[1]);
-        description += `${meat[1]} (±${Math.floor(meatInt * 0.2)}) meat\n`;
-      }
-      const dedupedDrops = new Map<string, number>();
-      let dropArray = [];
-      for (let drop of this._monster.drops) {
-        let dropDesc = "";
-        if (drop.attributes.accordion) {
-          dropDesc += `${hyperlink(drop.item, toWikiLink(drop.item))} (Stealable accordion)\n`;
-        } else {
-          dropDesc += `${hyperlink(drop.item, toWikiLink(drop.item))} (${
-            drop.droprate > 0 ? `${drop.droprate}%` : "Sometimes"
-          }`;
-          if (drop.attributes.pickpocketOnly) dropDesc += ", pickpocket only";
-          if (drop.attributes.noPickpocket) dropDesc += ", can't be stolen";
-          if (drop.attributes.conditional) dropDesc += ", conditional";
-          if (drop.attributes.fixedRate) dropDesc += ", unaffected by item drop modifiers";
-          dropDesc += ")";
-        }
-        if (dedupedDrops.has(dropDesc))
-          dedupedDrops.set(dropDesc, (dedupedDrops.get(dropDesc) || 0) + 1);
-        else dedupedDrops.set(dropDesc, 1);
-        dropArray.push(dropDesc);
-      }
-      let dropsWritten = 0;
-      for (let drop of dropArray) {
-        if (dedupedDrops.has(drop)) {
-          const quantity = dedupedDrops.get(drop) || 1;
-          description += `${quantity > 1 ? `${quantity}x ` : ""}${drop}\n`;
-          dedupedDrops.delete(drop);
-          dropsWritten += 1;
-          if (dropsWritten >= 10) break;
-        }
-      }
-      if (dedupedDrops.size > 0) description += "...and more.";
-    }
-    return description;
-  }
-
-  async addToEmbed(embed: EmbedBuilder, client: KoLClient): Promise<void> {
-    embed.setThumbnail(
-      `http://images.kingdomofloathing.com/adventureimages/${this._monster.imageUrl}`
-    );
-    if (!this._description) this._description = await this.buildDescription(client);
-    embed.setDescription(this._description);
-  }
-
-  parseMonsterData(monsterData: string): MonsterData {
-    const data = monsterData.split(/\t/);
-    if (data.length < 4) throw "Invalid data";
-    return {
-      name: cleanString(data[0]),
-      id: parseInt(data[1]),
-      imageUrl: data[2].split(",")[0],
-      parameters: data[3],
-      drops: data[4]
-        ? (
-            data
-              .slice(4)
-              .map((drop) => convertToDrop(cleanString(drop)))
-              .filter((drop) => drop !== undefined) as Drop[]
-          ).sort((a, b) => b.droprate - a.droprate)
-        : [],
-    };
-  }
-}
